+++ src/main/java/de/hhbk/jahresprojekt/model/Role.java
@@ -3,20 +3,20 @@ package de.hhbk.jahresprojekt.model;
 import java.util.List;
 
 public class Role {
-    private int id;
+    private Long id;
     private RoleType roleType;
     private List<User> userList;
 
     public Role() {
     }
 
-    public Role(int id, RoleType roleType, List<User> userList) {
+    public Role(Long id, RoleType roleType, List<User> userList) {
         this.id = id;
         this.roleType = roleType;
         this.userList = userList;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
